♻️ refactor: consolidate UI style files
